The solution
In this project I created a a solution to the 8 queen problem using simple java-code. The 8 queen problem revolves around combinations in
which you can place 8 queens on a chessboard without them threatening one another. There are in total 92 solutions. This application lets
you write a a chess-coordinate into the terminal (using algeabraic notation - i.e a character and number combination from A-H and 1-8). The
application will then display a one of the solutions to you in the terminal with a queen placed in the coordinate you typed in